Remove cast to int to be able to set float values

+++ custom_components/luxtronik/number.py
@@ -147,7 +147,7 @@ class LuxtronikNumberEntity(LuxtronikEntity, NumberEntity):
         value = self._pending_value
 
         if self.entity_description.factor is not None:
-            value = int(value / self.entity_description.factor)
+            value = value / self.entity_description.factor
         data = await self.coordinator.async_write(
             self.entity_description.luxtronik_key.value.split(".")[1], value
         )
